Ejercicios 7 y 9 modificados

+++ 1-EntradaSalida/jsEntradaSalida-09.js
@@ -6,8 +6,15 @@ en el cuadro de texto "RESULTADO".*/
 function mostrarAumento()
 {
 	let importe; let resAumento;
-	importe = parseInt(document.getElementById("txtIdSueldo").value);
+	//podemos usar parseFloat ya que un salario puede tener numeros decimales.
+	//Hay una forma para que me muestre solamente los numeros que yo necesite como decimales.
+	//ejemplo 2 numeros despues de la coma y no mas.
+	//importe = parseInt(document.getElementById("txtIdSueldo").value);
+
+	importe = parseFloat(document.getElementById("txtIdSueldo").value);
 	// resAumento = importe * 1.1;
+	// resAumento = importe * 0.1;
+	
 	resAumento = importe + ((importe*10)/100);
 	alert(document.getElementById("txtIdResultado").value = resAumento);
 }
